python/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mr_calibration"
version = "0.1.0"
authors = [
  { name="Alex Arbogast", email="arbogastaw@gmail.com" },
]
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
  "numpy<2",
  "pandas",
  "matplotlib",
  "roboticstoolbox-python",
  "scipy",
  "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.urls]
homepage = "https://github.com/Georgia-Tech-Manufacturing-Institute/multi_robot_calibration"
documentation = "https://github.com/Georgia-Tech-Manufacturing-Institute/multi_robot_calibration"
issues = "https://github.com/Georgia-Tech-Manufacturing-Institute/multi_robot_calibration/issues"

[project.scripts]
calibrate-bf = "mr_calibration.bf_calibration:main"
calibrate-iterative = "mr_calibration.iterative_calibration:main"
calibrate-tool = "mr_calibration.tool_calibration:main"
calibrate-wobj = "mr_calibration.wobj_calibration:main"

[tool.setuptools.package-data]
packages = ["data/*.csv"]
